+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 _No unreleased changes_
 
 ## [3.2.0] - 2025-07-06
+### Added 
+- Added twitch rewards queue link
+
 ### Changed
 - Changed channel point redemptions to use EventSub